Added
This is a console program that will calculate all possible combinations of coins (that is, quarters, dimes, nickels, and pennies) for a given amount of money and displays them in a little chart. I developed this project after my friend was having issues figuring out how the different combinations for a certain amount of money (it was for an school assignment), and I got to thinking about how I could write a program that was generalized to solve for any amount you gave it. The main block of logic for solving the problem is a little bit of a headache to wrap your mind around, but it was a blast to think through and code for.