Added
PNFO stands for "Prime Number Figure-Outer" (creative, I know). This program will determine if a given number is prime. I was going through some college courses a while back, a few of which were math classes, and I would quite regularly need to know if a number was prime or not. I could figure it out with enough time, but in order to speed things up, I developed this little application. It tells you whether the number is prime, but that's it. I later decided I wanted this program to do more, so I made second and third versions with increasingly greater functionality. Check those out too!