Added
This is the second version of my "Prime Number Figure-Outer." In version 2, I added the ability to check multiple numbers before the program exits. It will ask for a number, solve for it, and will ask if the user wants to solve another number. This is the second of three versions of this program. Check out the other two if you haven't!