Added
This is the third version of my "Prime Number Figure-Outer." In this version, I made the program keep track of and display the different factors of a number. Often times if you want to know if a number is prime or not, you will also want to know the factors of that number if it is not prime. So I added that functionality on top of the  previous functionality to help me in my math courses. This is version three, go and take a look at versions one and two if you haven't seen them already!