Added
This is my first attempt at creating the game "Connect 4" in Python. I had just gone through my python course at the time, and I was a little green when I developed this. Looking back I would have done a bunch differently (like not using Turtle graphics as the display), but I am proud of what I accomplished even if it was not all done according to best practices. I consider it a stepping stone in my programming journey, and I had a blast making it.