Added
This is my second attempt at making the game "Connect 4" in python. I attempted to write this a little later on. I was not super pleased with my first attempt once I sat back and thought about all the hacks I had to do to get it to work right (mainly thinking about how I used the "Turtle" graphics library). Because of that, I endeavored to create version 2 using the more capable "pygame" library. It was made for developing games, and so I thought it was the most suitable choice to use when making Connect 4. I'm still not entirely pleased with it (I had intended to make a scaling mechanism to change the size of the board relatively easily, but didn't) but even so, I am still proud of the work and learning that went into making this version.