Added
This is a little ASCII art program that prints out a rocket ship. I actually initially did this program for my Java programming course, but then I wanted to implement it in Python (just to say I did it in Python). The coolest thing about this program, I'd say, is that you can change one constant and the output gets scaled up or down accordingly.